Python/curso/Piscina_Python/ejercicios/Python_Piscina_Definitiva/requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==2.2.3
Pillow==9.0.1
Pillow==11.0.0
setuptools==59.6.0
